refactor(RacingTest): Separate RacingTest into normal and exception output
